Merge branch 'dev' of github.com:Qubinode/qubinode-installer

+++ lib/qubinode_utils.sh
@@ -13,24 +13,27 @@ function qubinode_project_cleanup () {
         FILES=("${FILES[@]}" "$vault_vars_file" "$vars_file")
     fi
 
-    # Delete OpenShift files
-    openshift_product=$(awk '/^product:/ {print $2}' "${project_dir}/playbooks/vars/all.yml")
-    if [[ ${openshift_product} == "ocp3" ]]; then
-      FILES=("${FILES[@]}" "$ocp3_vars_files")
-    elif [[ ${openshift_product} == "okd3" ]]; then
-      FILES=("${FILES[@]}" "$okd3_vars_files")
-    fi
+    # Delete OpenShift 3 files
+    if [ -f ${project_dir}/playbooks/vars/ocp3.yml ]
+    then 
+        openshift_product=$(awk '/^product:/ {print $2}' "${project_dir}/playbooks/vars/ocp3.yml")
+        if [[ ${openshift_product} == "ocp3" ]]; then
+          FILES=("${FILES[@]}" "$ocp3_vars_files")
+        elif [[ ${openshift_product} == "okd3" ]]; then
+          FILES=("${FILES[@]}" "$okd3_vars_files")
+        fi
 
-    if [ ${#FILES[@]} -eq 0 ]
-    then
-        echo "Project directory: ${project_dir} state is already clean"
-    else
-        for f in $(echo "${FILES[@]}")
-        do
-            test -f $f && rm $f
-            echo "purged $f"
+        if [ ${#FILES[@]} -eq 0 ]
+        then
+            echo "Project directory: ${project_dir} state is already clean"
+        else
+            for f in $(echo "${FILES[@]}")
+            do
+                test -f $f && rm $f
+                echo "purged $f"
 
-        done
+            done
+        fi
     fi
 
    echo "Removing playbook vars"
@@ -64,13 +67,13 @@ function canSSH () {
 
 
 function get_admin_user_password () {
-    echo "Fetching OpenShift Admin Password. Please Enter Vault password to decrypt file."
-    ansible-vault decrypt "${vault_vars_file}"
+    #echo " Fetching the Admin user password."
+    decrypt_ansible_vault "${vault_vars_file}" > /dev/null
     admin_user_passowrd=$(awk '/admin_user_password:/ {print $2}' "${vault_vars_file}")
-    ansible-vault encrypt "${vault_vars_file}"
+    encrypt_ansible_vault "${vaultfile}" >/dev/null
     if [ "A${admin_user_passowrd}" == "A" ]
     then
-        echo "Unable to retrieve $CURRENT_USER user password from the vault"
+        print "%s\n" " Unable to retrieve ${yel}$CURRENT_USER${end} user password from the vault"
         exit 1
     fi
 }
@@ -85,3 +88,26 @@ function exit_status () {
         exit "${RESULT}"
     fi
 }
+
+
+convertB_human() {
+    # Thanks to https://bit.ly/39xomtN
+    NUMBER=$1
+    for DESIG in Bytes KB MB GB TB PB
+    do
+        [ $NUMBER -lt 1024 ] && break
+        let NUMBER=$NUMBER/1024
+    done
+
+    DISK_SIZE_HUMAN=$(printf "%d %s\n" $NUMBER $DESIG)
+    DISK_SIZE_COMPARE=$(printf "%d %s\n" $NUMBER)
+}
+
+function isvmRunning () {
+    sudo virsh list |grep $vm|awk '/running/ {print $2}'
+}
+
+function isvmShutdown () {
+    sudo virsh list --all | grep $vm| awk '/shut/ {print $2}'
+}
+
